pyproject.toml
[tool.poetry]
name = "cinepy"
version = "0.1.0"
description = ""
authors = ["HenriqueHartmann <alemaohartmann@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.1"
dj-database-url = "^0.5.0"
environs = "^9.3.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
